20240513/1/pyproject.toml
[project]
name = "Mood"
version = "1.0.0"
dependencies = ["python-cowsay"]
description = "Moo! Monster! Moo! Spooky! Moo-oo-o!"

[build-system]
requires = ['setuptools']
build-backend = "setuptools.build_meta"

[tool.setuptools.package-dir]
mood = "mood"

[tool.setuptools.package-data]
mood = ["mood/server/*/*/*.mo", "mood/docs/*"]